requirements.txt
pydantic>=2.0.0
typing-extensions>=4.0.0
requests>=2.25.0
python-dotenv>=1.0.0
rich>=10.0.0
setuptools>=42.0.0
flask>=2.0.0
flask-cors>=3.0.0
